split qos aware throwables
