+++ atlasdb-commons/src/main/java/com/palantir/common/base/Throwables.java
@@ -29,7 +29,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.google.common.base.Preconditions;
-import com.palantir.atlasdb.qos.ratelimit.RateLimitExceededException;
 import com.palantir.common.exception.AtlasDbDependencyException;
 import com.palantir.common.exception.PalantirRuntimeException;
 
@@ -78,22 +77,6 @@ public final class Throwables {
         rewrapAndThrowIfInstance(newMessage, ex, RuntimeException.class);
         rewrapAndThrowIfInstance(newMessage, ex, Error.class);
         throw createPalantirRuntimeException(newMessage, ex);
-    }
-
-    /**
-     * If the provided Throwable is
-     *   a) an ExecutionException or InvocationTargetException, then apply this method on the cause;
-     *   b) a RateLimitExceededException or an AtlasDbDependencyException, then rethrow it;
-     *   c) none of the above, then wrap it in an AtlasDbDependencyException and throw that.
-     */
-    public static RuntimeException unwrapAndThrowRateLimitExceededOrAtlasDbDependencyException(Throwable ex) {
-        if (ex instanceof ExecutionException || ex instanceof InvocationTargetException) {
-            // Needs to be this way in case you have ITE(RLE) or variants of that.
-            throw unwrapAndThrowRateLimitExceededOrAtlasDbDependencyException(ex.getCause());
-        } else if (ex instanceof RateLimitExceededException) {
-            throw (RateLimitExceededException) ex;
-        }
-        throw createAtlasDbDependencyException(ex);
     }
 
     /**
