+++ atlasdb-commons/src/test/java/com/palantir/common/base/ThrowablesTest.java
@@ -15,12 +15,8 @@
  */
 package com.palantir.common.base;
 
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
 import java.io.IOException;
-import java.lang.reflect.InvocationTargetException;
 import java.sql.SQLException;
-import java.util.concurrent.ExecutionException;
 
 import org.junit.Assert;
 import org.junit.Before;
@@ -30,11 +26,6 @@ import com.palantir.atlasdb.qos.ratelimit.RateLimitExceededException;
 import com.palantir.common.exception.AtlasDbDependencyException;
 
 public class ThrowablesTest extends Assert {
-    private static final Exception RATE_LIMIT_EXCEEDED_EXCEPTION =
-            new RateLimitExceededException("Stop!");
-    private static final Exception ATLASDB_DEPENDENCY_EXCEPTION =
-            new AtlasDbDependencyException("The TimeLock is dead, long live the TimeLock");
-
     @Before
     public void setUp() throws Exception {
         NoUsefulConstructorException.noUsefulConstructorCalled = false;
@@ -73,45 +64,6 @@ public class ThrowablesTest extends Assert {
         }
     }
 
-    @Test
-    public void unwrapAndThrowRateLimitExceededOrAtlasDbDependencyExceptionCanThrowRateLimitExceededException() {
-        assertThatThrownBy(() -> Throwables.unwrapAndThrowRateLimitExceededOrAtlasDbDependencyException(
-                RATE_LIMIT_EXCEEDED_EXCEPTION)).isEqualTo(RATE_LIMIT_EXCEEDED_EXCEPTION);
-    }
-
-    @Test
-    public void unwrapAndThrowRateLimitExceededOrAtlasDbDependencyExceptionCanThrowAtlasDbDependencyException() {
-        assertThatThrownBy(() -> Throwables.unwrapAndThrowRateLimitExceededOrAtlasDbDependencyException(
-                ATLASDB_DEPENDENCY_EXCEPTION)).isEqualTo(ATLASDB_DEPENDENCY_EXCEPTION);
-    }
-
-    @Test
-    public void unwrapAndThrowRateLimitExceededOrAtlasDbDependencyExceptionThrowsWrappedRateLimitExceededExceptions() {
-        assertThatThrownBy(() -> Throwables.unwrapAndThrowRateLimitExceededOrAtlasDbDependencyException(
-                new ExecutionException(RATE_LIMIT_EXCEEDED_EXCEPTION))).isEqualTo(RATE_LIMIT_EXCEEDED_EXCEPTION);
-        assertThatThrownBy(() -> Throwables.unwrapAndThrowRateLimitExceededOrAtlasDbDependencyException(
-                new InvocationTargetException(RATE_LIMIT_EXCEEDED_EXCEPTION))).isEqualTo(RATE_LIMIT_EXCEEDED_EXCEPTION);
-    }
-
-    @Test
-    public void unwrapAndThrowRateLimitExceededOrAtlasDbDependencyExceptionThrowsWrappedAtlasDbDependencyExceptions() {
-        assertThatThrownBy(() -> Throwables.unwrapAndThrowRateLimitExceededOrAtlasDbDependencyException(
-                new ExecutionException(ATLASDB_DEPENDENCY_EXCEPTION))).isEqualTo(ATLASDB_DEPENDENCY_EXCEPTION);
-        assertThatThrownBy(() -> Throwables.unwrapAndThrowRateLimitExceededOrAtlasDbDependencyException(
-                new InvocationTargetException(ATLASDB_DEPENDENCY_EXCEPTION))).isEqualTo(ATLASDB_DEPENDENCY_EXCEPTION);
-    }
-
-    @Test
-    public void unwrapAndThrowRateLimitExceededOrAtlasDbDependencyExceptionWrapsRuntimeExceptions() {
-        assertThatThrownBy(() -> Throwables.unwrapAndThrowRateLimitExceededOrAtlasDbDependencyException(
-                new RuntimeException("runtimeException"))).isInstanceOf(AtlasDbDependencyException.class);
-    }
-
-    @Test
-    public void unwrapAndThrowRateLimitExceededOrAtlasDbDependencyExceptionWrapsCheckedExceptions() {
-        assertThatThrownBy(() -> Throwables.unwrapAndThrowRateLimitExceededOrAtlasDbDependencyException(
-                new IOException("ioException"))).isInstanceOf(AtlasDbDependencyException.class);
-    }
 
     // only has a (string, throwable) constructor
     public void throwTwoArgConstructorException() throws TwoArgConstructorException {
